backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
mysql-connector-python==8.1.0
PyJWT==2.8.0
python-dotenv==1.0.0
flask-swagger-ui==3.18.0
openpyxl==3.1.2